Create SearchMotifs.m
The first function searches the data for motifs that are known from the PSP database to be phosphorylated by a certain kinase. The database can be downloaded from http://www.phosphosite.org/staticDownloads.do. It should be reduced to one column containing the kinase names and one containing the motifs. Since the exact position within the motif is of great importance, the function compares the strings which are left and right from the central amino acid to those of the motifs in the database. It adds a column with the kinases found to the data matrix and takes as an input the original dataset as a cell array, the PSP database as a two-column cell array and the column containing the motif in the dataset. The data matrix should begin with a column containing the protein IDs (e.g. gene names) followed by a column for the motifs and a variable number of data columns. The phosphorylated amino acid should be in the middle of the motif.