Create MakeKin.m
This function creates a list of the kinases that are known to phosphorylate at least one motif in the dataset. If additionally proteome data are available (optional 3rd argument, list of protein names) the function reduces the kinase list to those which are present in the proteome. Finally, it produces a new data matrix which only contains those phosphosites for which a kinase is known.