Create CreateTT.m
The third function creates a truth table out of the reduced dataset, with 1 for true interactions and 0 for false. This will be used as a mapping in the next step to estimate the kinase activities and affinity parameters on the basis of only those measurement values, the respective kinases have an influence on.