Create FitActivities.m
FitActivities estimates the activities of the kinases in kin using the Matlab built-in function fmincon. The number of iterations can be specified using the input argument iter. Each iteration starts with randomized parameter vectors and cycles between an optimization of the affinity parameters by means of the global cost function and an optimization of the kinase activities for each condition by means of the local cost function. Since the gradients for the parameters are delivered by the cost functions, the ‘GradObj’ argument in the option settings should be turned to ‘on’. The while-loop continues as long as the difference between both costs is larger than a certain value, in this case 10. The upper and lower bounds for parameter estimation as well as the starting values should be modified according to the user´s needs. The output variables AP and K contain the estimated affinity parameters and kinase activities, cost is a list of all calculated costs with mincost being the lowest. This function can also be executed on multiple cores using the parallel computing toolbox. In this case, the first for-loop needs to be replaced by a parfor-loop.